fix: V3.4 — deadlock fix, watchdog tuning, OLED optimization, docs refresh
- Bug #1 (致命): s_net_state not reset to NET_IDLE after disconnect/silent
  timeout, causing permanent deadlock on KEY0 reconnect. Added
  s_net_state=NET_IDLE in both silent watchdog and CLOSED paths.
- Bug #2: silent watchdog 15s→30s (too aggressive for manual ops).
  Added ESP8266_RefreshLastRxTime() called at WiFi connect success
  to give fresh 30s grace window. Fixed TCP keepalive misinformation.
- Bug #3: NetAssist guide referenced main.c for WiFi config; now
  correctly points to App_Net.c (where macros actually live).
- Bug #4: ESP8266_CopyRxFrame() now consumes \r\n as a pair,
  eliminating empty newline residual frames on next poll cycle.
- Simplify: OLED ShowNum/ShowSignedNum/ShowHexNum/ShowBinNum
  precompute divisor (O(n²)→O(n)). ShowFloatNum double→float.
  OLED_Init delay loop volatile to prevent compiler elimination.
  PWM.h comment fix (20ms→10ms/step). WiFi LED now solid on connect.
- Docs: architecture guide V3.3→V3.4, NetAssist guide v1.2→v1.3,
  all .docx regenerated. README badges and matrix synced.

+++ Keil_Project/Hardware/OLED.h
@@ -20,6 +20,6 @@ void OLED_ShowNum(uint8_t Line, uint8_t Column, uint32_t Number, uint8_t Length)
 void OLED_ShowSignedNum(uint8_t Line, uint8_t Column, int32_t Number, uint8_t Length);
 void OLED_ShowHexNum(uint8_t Line, uint8_t Column, uint32_t Number, uint8_t Length);
 void OLED_ShowBinNum(uint8_t Line, uint8_t Column, uint32_t Number, uint8_t Length);
-void OLED_ShowFloatNum(uint8_t Line, uint8_t Column, double Number, uint8_t IntLength, uint8_t FractLength);
+void OLED_ShowFloatNum(uint8_t Line, uint8_t Column, float Number, uint8_t IntLength, uint8_t FractLength);
 
 #endif
